+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# vNext
+
+- (bug) Cleanup tests
+
 # v4.0.0-alpha.2
 
 - (bug) correct version tagging